+++ tests/resources/components/thread/comment.blade.php
@@ -1,0 +1,9 @@
+@props(['comment'])
+
+<li>
+    <span>{{ $comment['message'] }}</span>
+
+    @if (isset($comment['comments']) && count($comment['comments']) > 0)
+        <c-thread :comments="$comment['comments']" />
+    @endif
+</li>