+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Unreleased
 
 - Adds a new "Compile Time Rendering" system, which can render components at compile time and inline the static output
+- Adds compiler support for circular component references, like nested comment threads
+- Adds a `#cache` compiler attribute, which may be used to cache the results of any Dagger component
+- Bumps the minimum Laravel version to `11.23`, for `Cache::flexible` support
+- Improves compilation of custom functions declared within a component's template
 
 ## [v1.0.6](https://github.com/Stillat/dagger/compare/v1.0.5...v1.0.6) - 2025-01-31